refactor(aapd-861): split and tidy integration tests
