+++ packages/appeals-service-api/__tests__/developer/back-office.test.js
@@ -2,27 +2,29 @@ const http = require('http');
 const supertest = require('supertest');
 const { MongoClient } = require('mongodb');
 const container = require('rhea');
-const uuid = require('uuid');
-const crypto = require('crypto');
 const jp = require('jsonpath');
-// const logger = require('../../src/lib/logger');
 
 const app = require('../../src/app');
 const appDbConnection = require('../../src/db/db');
 const appConfiguration = require('../../src/configuration/config');
+
 const MockedExternalApis = require('./external-dependencies/rest-apis/mocked-external-apis');
 const HorizonInteraction = require('./external-dependencies/rest-apis/interactions/horizon-interaction');
 const NotifyInteraction = require('./external-dependencies/rest-apis/interactions/notify-interaction');
 const AppealFixtures = require('./fixtures/appeals');
-const FinalCommentFixtures = require('./fixtures/finalComments');
 const HorizonIntegrationInputCondition = require('./utils/horizon-integration-input-condition');
 
 const { isFeatureActive } = require('../../src/configuration/featureFlag');
 
+/** @type {import('supertest').SuperTest<import('supertest').Test>} */
 let appealsApi;
+/** @type {import('mongodb').MongoClient} */
 let databaseConnection;
+/** @type {import('./external-dependencies/rest-apis/mocked-external-apis')} */
 let mockedExternalApis;
+/** @type {Array.<*>} */
 let expectedHorizonInteractions;
+/** @type {Array.<*>} */
 let expectedNotifyInteractions;
 let testLpaEmail = 'appealplanningdecisiontest@planninginspectorate.gov.uk';
 let testLpaCodeEngland = 'E69999999';
@@ -35,6 +37,8 @@ jest.setTimeout(240000); // The Horizon integration tests need a bit of time to 
 jest.mock('../../src/db/db'); // TODO: We shouldn't need to do this, but we didn't have time to look at making this better. It should be possible to use the DB connection directly (not mock it)
 jest.mock('../../src/configuration/featureFlag');
 
+const dbName = 'integration-test-back-office-db';
+
 beforeAll(async () => {
 	///////////////////////////////
 	///// SETUP EXTERNAL APIs /////
@@ -45,12 +49,15 @@ beforeAll(async () => {
 	///////////////////////////////
 	///// SETUP TEST DATABASE /////
 	///////////////////////////////
+	if (!process.env.INTEGRATION_TEST_DB_URL) {
+		throw new Error('process.env.INTEGRATION_TEST_DB_URL not set');
+	}
 
 	databaseConnection = await MongoClient.connect(process.env.INTEGRATION_TEST_DB_URL, {
 		useNewUrlParser: true,
 		useUnifiedTopology: true
 	});
-	let mockedDatabase = await databaseConnection.db('foo');
+	let mockedDatabase = await databaseConnection.db(dbName);
 	appDbConnection.get.mockReturnValue(mockedDatabase);
 
 	const test_listener = container.create_container();
@@ -62,14 +69,6 @@ beforeAll(async () => {
 	/////////////////////////////
 	///// SETUP TEST CONFIG /////
 	/////////////////////////////
-
-	appConfiguration.secureCodes.finalComments.length = 4;
-	appConfiguration.secureCodes.finalComments.expirationTimeInMinutes = 30;
-	appConfiguration.secureCodes.finalComments.decipher.algorithm = 'aes-256-cbc';
-	appConfiguration.secureCodes.finalComments.decipher.initVector = crypto.randomBytes(16);
-	appConfiguration.secureCodes.finalComments.decipher.securityKey = crypto.randomBytes(32);
-	appConfiguration.secureCodes.finalComments.decipher.inputEncoding = 'hex';
-	appConfiguration.secureCodes.finalComments.decipher.outputEncoding = 'utf-8';
 	appConfiguration.services.horizon.url = mockedExternalApis.getHorizonUrl();
 	appConfiguration.services.notify.apiKey =
 		'hasserviceapikey-u89q754j-s87j-1n35-s351-789245as890k-1545v789-8s79-0124-qwe7-j2vfds34w5nm';
@@ -126,78 +125,6 @@ afterEach(async () => {
 afterAll(async () => {
 	await databaseConnection.close();
 	await mockedExternalApis.teardown();
-});
-
-describe('Appeals', () => {
-	it(`should return an error if we try to update an appeal that doesn't exist`, async () => {
-		// When: an appeal is sent via a PUT or PATCH request, but hasn't yet been created
-		const householderAppeal = AppealFixtures.newHouseholderAppeal(uuid.v4());
-		const putResponse = await appealsApi
-			.put(`/api/v1/appeals/${householderAppeal.id}`)
-			.send(householderAppeal);
-
-		const patchResponse = await appealsApi
-			.patch(`/api/v1/appeals/${householderAppeal.id}`)
-			.send(householderAppeal);
-
-		// Then: we should get a 404 status code for both requests
-		expect(putResponse.status).toBe(404);
-		expect(patchResponse.status).toBe(404);
-
-		// And: external systems should be interacted with in the following ways
-		expectedHorizonInteractions = [];
-		expectedNotifyInteractions = [];
-	});
-
-	it("should apply patch updates correctly when data to patch-in isn't a full appeal", async () => {
-		// Given: an appeal is created
-		const savedAppealResponse = await _createAppeal();
-		let savedAppeal = savedAppealResponse.body;
-
-		// When: the appeal is patched
-		const patchedAppealResponse = await appealsApi
-			.patch(`/api/v1/appeals/${savedAppeal.id}`)
-			.send({ horizonId: 'foo' });
-
-		// Then: when we retrieve the appeal, it should have the patch applied
-		savedAppeal.horizonId = 'foo';
-		savedAppeal.updatedAt = patchedAppealResponse.body.updatedAt;
-		expect(patchedAppealResponse.body).toMatchObject(savedAppeal);
-
-		// And: no external systems should be interacted with
-		expectedHorizonInteractions = [];
-		expectedNotifyInteractions = [];
-	});
-
-	it('should return the relevant appeal when requested after the appeal has been saved', async () => {
-		// Given: an appeal is created
-		const savedAppeal = await _createAppeal();
-
-		// When: we try to request that appeal
-		const requestedAppeal = await appealsApi.get(`/api/v1/appeals/${savedAppeal.body.id}`);
-
-		// Then: we should get a 200 status
-		expect(requestedAppeal.status).toEqual(200);
-
-		// And: the correct appeal should be returned
-		expect(requestedAppeal.body.id).toEqual(savedAppeal.body.id);
-
-		// And: external systems should be interacted with in the following ways
-		expectedHorizonInteractions = [];
-		expectedNotifyInteractions = [];
-	});
-
-	it(`should return an error if an appeal is requested that doesn't exist`, async () => {
-		// When: we try to access a non-existent appeal
-		const getAppealResponse = await appealsApi.get(`/api/v1/appeals/${uuid.v4()}`);
-
-		// Then: we should get a 404 status
-		expect(getAppealResponse.status).toEqual(404);
-
-		// And: external systems should be interacted with in the following ways
-		expectedHorizonInteractions = [];
-		expectedNotifyInteractions = [];
-	});
 });
 
 describe('Back Office', () => {
@@ -1733,161 +1660,6 @@ describe('Back Office', () => {
 	});
 });
 
-describe('Final comments', () => {
-	it('Should submit the final comment to the back-end when submitted before the expiry date', async () => {
-		// Given: a final comment with an expiry date of tomorrow
-		let today = new Date();
-		let tomorrow = new Date();
-		tomorrow.setDate(today.getDate() + 1);
-		let expiryDate = tomorrow.toISOString();
-		let horizonId = 1345678;
-
-		let finalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: horizonId
-		});
-
-		// When: it is submitted to the back-end
-		let response = await _createFinalComment(finalCommentToSubmit);
-
-		// Then: It should return a status code of 201
-		expect(response.status).toBe(201);
-
-		// And: The back-end should contain one appeal for that horizon Id
-		let submittedFinalComment = await appealsApi.get(`/api/v1/final-comments/${horizonId}`);
-		expect(submittedFinalComment.status).toBe(201);
-		expect(submittedFinalComment.body.length).toBe(1);
-	});
-
-	it('Should not submit the final comment if a final comment exists with the same horizonId and email', async () => {
-		// Given: a final comment with an expiry date of tomorrow
-		let today = new Date();
-		let tomorrow = new Date();
-		tomorrow.setDate(today.getDate() + 1);
-		let expiryDate = tomorrow.toISOString();
-		let horizonId = 1345678;
-
-		let finalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: horizonId
-		});
-
-		// When: it is submitted to the back-end twice
-		let firstResponse = await _createFinalComment(finalCommentToSubmit);
-		let secondResponse = await _createFinalComment(finalCommentToSubmit);
-
-		// Then: It should return a status code of 201 for the first
-		expect(firstResponse.status).toBe(201);
-
-		// And: Return a status code of 409 for the second
-		expect(secondResponse.status).toBe(409);
-
-		// And: The back-end should contain one final comment for that horizon Id
-		let submittedFinalComment = await appealsApi.get(`/api/v1/final-comments/${horizonId}`);
-		expect(submittedFinalComment.status).toBe(201);
-		expect(submittedFinalComment.body.length).toBe(1);
-	});
-
-	it('Should not submit the final comment if the expiry date is in the past', async () => {
-		// Given: A final comment with an expiry date in the past
-		let today = new Date();
-		let yesterday = new Date();
-		yesterday.setDate(today.getDate() - 1);
-		let expiryDate = yesterday.toISOString();
-		let horizonId = 1345678;
-
-		let finalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: horizonId
-		});
-
-		// When: it is submitted to the back end
-		let response = await _createFinalComment(finalCommentToSubmit);
-
-		// Then: It should return a status code of 409
-		expect(response.status).toBe(409);
-
-		// And: The back-end should not contain an final comment for the horizon Id
-		let submittedFinalComment = await appealsApi.get(`/api/v1/final-comments/${horizonId}`);
-		expect(submittedFinalComment.status).toBe(404);
-	});
-
-	it('Should submit 2 final comments if the final comments have a different horizonId', async () => {
-		// Given: a final comment with an expiry date of tomorrow
-		let today = new Date();
-		let tomorrow = new Date();
-		tomorrow.setDate(today.getDate() + 1);
-		let expiryDate = tomorrow.toISOString();
-		let firstHorizonId = 1345678;
-		let secondHorizonId = 9999999;
-
-		let firstFinalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: firstHorizonId
-		});
-
-		let secondFinalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: secondHorizonId
-		});
-
-		// When: it is submitted to the back-end twice
-		let firstResponse = await _createFinalComment(firstFinalCommentToSubmit);
-		let secondResponse = await _createFinalComment(secondFinalCommentToSubmit);
-
-		// Then: It should return a status code of 201 for the both
-		expect(firstResponse.status).toBe(201);
-		expect(secondResponse.status).toBe(201);
-
-		// And: The back-end should contain one final comment for each horizon Id
-		let firstSubmittedFinalComment = await appealsApi.get(
-			`/api/v1/final-comments/${firstHorizonId}`
-		);
-		expect(firstSubmittedFinalComment.status).toBe(201);
-		expect(firstSubmittedFinalComment.body.length).toBe(1);
-
-		let secondSubmittedFinalComment = await appealsApi.get(
-			`/api/v1/final-comments/${secondHorizonId}`
-		);
-		expect(secondSubmittedFinalComment.status).toBe(201);
-		expect(secondSubmittedFinalComment.body.length).toBe(1);
-	});
-
-	it('Should submit 2 final comments if the final comments have the same horizonId but different emails', async () => {
-		// Given: a final comment with an expiry date of tomorrow
-		let today = new Date();
-		let tomorrow = new Date();
-		tomorrow.setDate(today.getDate() + 1);
-		let expiryDate = tomorrow.toISOString();
-		let horizonId = 9999999;
-
-		let firstFinalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: horizonId,
-			email: 'test@example.com'
-		});
-
-		let secondFinalCommentToSubmit = FinalCommentFixtures.newFinalComment({
-			finalCommentExpiryDate: expiryDate,
-			horizonId: horizonId,
-			email: 'test2@example.com'
-		});
-
-		// When: it is submitted to the back-end twice
-		let firstResponse = await _createFinalComment(firstFinalCommentToSubmit);
-		let secondResponse = await _createFinalComment(secondFinalCommentToSubmit);
-
-		// Then: It should return a status code of 201 for the both
-		expect(firstResponse.status).toBe(201);
-		expect(secondResponse.status).toBe(201);
-
-		// And: The back-end should contain two final comments for that horizonId
-		let submittedFinalComments = await appealsApi.get(`/api/v1/final-comments/${horizonId}`);
-		expect(submittedFinalComments.status).toBe(201);
-		expect(submittedFinalComments.body.length).toBe(2);
-	});
-});
-
 const _createAppeal = async (householderAppeal = AppealFixtures.newHouseholderAppeal()) => {
 	const appealCreatedResponse = await appealsApi.post('/api/v1/appeals');
 	const appealCreated = appealCreatedResponse.body;
@@ -1900,16 +1672,12 @@ const _createAppeal = async (householderAppeal = AppealFixtures.newHouseholderAp
 	return savedAppealResponse;
 };
 
-const _createFinalComment = async (finalComment) => {
-	return await appealsApi.post('/api/v1/final-comments').send(finalComment);
-};
-
 /**
  * Clears out all collection from the database EXCEPT the LPA collection, since this is needed
  * from one test to the next, and its data does not change during any test execution.
  */
 const _clearDatabaseCollections = async () => {
-	const databaseCollections = await databaseConnection.db('foo').collections();
+	const databaseCollections = await databaseConnection.db(dbName).collections();
 
 	const databaseCollectionsFiltered = databaseCollections.filter(
 		(collection) => collection.namespace.split('.')[1] !== 'lpa'
